refactor: split register constant classes into dedicated modules
Extract STATUS_RETURN, ERROR_BIT_MASKS, LED_STATE, and INSTRUCTION
from registers.py into their own files following the one-class-per-file
rule introduced in AGENTS.md.

registers.py now imports from these modules to preserve the existing
public API (registers.LED_STATE, registers.INSTRUCTION, etc.).
Adds shebang and encoding header to registers.py per new policy.
