+++ pydynamixel/led_state.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""LED state constants."""
+
+
+class LED_STATE:
+    OFF = 0
+    ON = 1
