+++ pydynamixel/status_return.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""Status return level constants."""
+
+
+class STATUS_RETURN:
+    NO_STATUS_PACKET = 0
+    RETURN_ONLY_FOR_READ = 1
+    RETURN_FOR_ALL_PACKETS = 2
